--- app/src/main/java/com/kuit/ourmenu/ui/onboarding/model/MealTimeState.kt
@@ -1,6 +1,0 @@
-package com.kuit.ourmenu.ui.onboarding.model
-
-data class MealTimeState(
-    val mealTime: String,
-    var selected: Boolean = false
-)
