+++ app/src/main/java/com/kuit/ourmenu/ui/signup/navigation/SignupNavigation.kt
@@ -6,7 +6,7 @@ import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
 import com.kuit.ourmenu.ui.navigator.Routes
-import com.kuit.ourmenu.ui.onboarding.viewmodel.SignupViewModel
+import com.kuit.ourmenu.ui.signup.viewmodel.SignupViewModel
 import com.kuit.ourmenu.ui.signup.screen.SignupEmailScreen
 import com.kuit.ourmenu.ui.signup.screen.SignupMealTimeScreen
 import com.kuit.ourmenu.ui.signup.screen.SignupPasswordScreen
